notice bei nichtexistenter ht datei raus..
git-svn-id: svn://svn.rexdev.de/addons/mysql_tools/trunk@24 b819eca2-34c4-4fe9-9cff-fb58e1037e43

+++ mysql_tools/functions/function.a895_commons.inc.php
@@ -338,8 +338,9 @@ function a895_htaccessDelete($ht_owner='all')
 
     default:
       $ht_file = $REX['ADDON'][$mypage]['ht_files'][$ht_owner];
-      if(!unlink($ht_file))
-        $ret = false;
+      if(file_exists($ht_file))
+        if(!unlink($ht_file))
+          $ret = false;
   }
   return $ret;
 }
